Split into HeroImageView and FaviconImageView
